fix model
